+++ R/lmer.R
@@ -16,6 +16,9 @@
 
 
 lmerExp = function(lmer.coef, family ="binomial", dec){
+  if (class(lmer.coef) == "numeric"){
+    lmer.coef = t(data.frame(lmer.coef))
+  }
   pv = 2*(1-pnorm(abs(lmer.coef[,3])))
   if (family == "binomial"){
     OR = paste(round(exp(lmer.coef[,1]), dec), " (", round(exp(lmer.coef[,1] - 1.96*exp(lmer.coef[,2])), dec), ",", round(exp(lmer.coef[,1] + 1.96*exp(lmer.coef[,2])), dec),")", sep="")
@@ -35,7 +38,7 @@ lmerExp = function(lmer.coef, family ="binomial", dec){
 
 
 
-#' @title lmer.display: table for "lmerMod" of "glmerMod" object (lme4 package)
+#' @title lmer.display: table for "lmerMod" or "glmerMod" object (lme4 package)
 #' @description Make mixed effect model results from "lmerMod" or "glmerMod" object (lme4 package)
 #' @param lmerMod.obj "lmerMod" or "glmerMod" object
 #' @param dec Decimal, Default: 2
@@ -59,17 +62,6 @@ lmerExp = function(lmer.coef, family ="binomial", dec){
 lmer.display = function(lmerMod.obj, dec = 2, ci.ranef = F){
   sl = summary(lmerMod.obj)
   fixef = sl$coefficients[-1,]
-  ranef = data.frame(sl$varcor)[,c(1,4)]
-  ranef.out = cbind(as.character(round(ranef[,2], dec)), matrix(NA, nrow(ranef), 3))
-  rownames(ranef.out) = ranef[,1]
-  if (ci.ranef){
-    ranef.ci = confint.merMod(lmerMod.obj, parm = 1:nrow(ranef), oldNames = F)^2
-    ranef.paste = paste(round(ranef$vcov, dec)," (", round(ranef.ci[,1], dec), ",", round(ranef.ci[,2], dec),")", sep="")
-    ranef.out = cbind(ranef.paste, matrix(NA, nrow(ranef), 3))
-    rownames(ranef.out) = ranef[,1]
-  }
-  ranef.na = rbind(rep(NA, 4), ranef.out)
-  rownames(ranef.na)[1] = "Random effects"
   
   mdata = lmerMod.obj@frame
   forms = as.character(sl$call[[2]])
@@ -88,21 +80,47 @@ lmer.display = function(lmerMod.obj, dec = 2, ci.ranef = F){
     unis2 = Reduce(rbind, unis)
     uni.res <- unis2[rownames(unis2) != "(Intercept)",]
   }
-  fix.all = cbind(lmerExp(uni.res, family=family.lmer, dec = dec), lmerExp(fixef, family=family.lmer, dec = dec))
-  family.label = colnames(fix.all)[1]
-  colnames(fix.all) = c(paste("crude ", family.label, ".(95%CI)",sep = ""), "crude P value", paste("adj. ", family.label, ".(95%CI)",sep = ""), "adj. P value")
+  
+  if(length(xf) == 1){
+    fix.all = lmerExp(uni.res, family=family.lmer, dec = dec)
+    family.label = colnames(fix.all)[1]
+    colnames(fix.all) = c(paste(family.label, "(95%CI)",sep = ""), "P value")
+    rownames(fix.all) = rownames(sl$coefficients)[-1]
+  } else{
+    fix.all = cbind(lmerExp(uni.res, family=family.lmer, dec = dec), lmerExp(fixef, family=family.lmer, dec = dec))
+    family.label = colnames(fix.all)[1]
+    colnames(fix.all) = c(paste("crude ", family.label, "(95%CI)",sep = ""), "crude P value", paste("adj. ", family.label, "(95%CI)",sep = ""), "adj. P value")
+    rownames(fix.all) = rownames(sl$coefficients)[-1]
+  }
+  
+  ranef = data.frame(sl$varcor)[,c(1,4)]
+  ranef.out = cbind(as.character(round(ranef[,2], dec)), matrix(NA, nrow(ranef), ncol(fix.all) - 1 ))
+  rownames(ranef.out) = ranef[,1]
+  if (ci.ranef){
+    ranef.ci = confint.merMod(lmerMod.obj, parm = 1:nrow(ranef), oldNames = F)^2
+    ranef.paste = paste(round(ranef$vcov, dec)," (", round(ranef.ci[,1], dec), ",", round(ranef.ci[,2], dec),")", sep="")
+    ranef.out = cbind(ranef.paste, matrix(NA, nrow(ranef), 3))
+    rownames(ranef.out) = ranef[,1]
+  }
+  
+  ranef.na = rbind(rep(NA, ncol(fix.all)), ranef.out)
+  rownames(ranef.na)[1] = "Random effects"
   colnames(ranef.na) = colnames(fix.all)
+  
   
   ## rownames
   fix.all.list = lapply(xf, function(x){fix.all[grepl(x, rownames(fix.all)),]})      
-  varnum.mfac = which(lapply(fix.all.list, length) > 4)
-  lapply(varnum.mfac, function(x){fix.all.list[[x]] <<- rbind(rep(NA, 4), fix.all.list[[x]])})
+  varnum.mfac = which(lapply(fix.all.list, length) > ncol(fix.all))
+  lapply(varnum.mfac, function(x){fix.all.list[[x]] <<- rbind(rep(NA, ncol(fix.all)), fix.all.list[[x]])})
   fix.all.unlist = Reduce(rbind, fix.all.list)
   
   rn.list = lapply(xf, function(x){rownames(fix.all)[grepl(x, rownames(fix.all))]})
   varnum.2fac = which(lapply(xf, function(x){length(sapply(mdata, levels)[[x]])}) == 2)
   lapply(varnum.2fac, function(x){rn.list[[x]] <<- paste(xf[x], ": ", levels(mdata[, xf[x]])[2], " vs ", levels(mdata[, xf[x]])[1], sep="")})
   lapply(varnum.mfac, function(x){rn.list[[x]] <<- c(paste(xf[x],": ref.=", levels(mdata[, xf[x]])[1], sep=""), gsub(xf[x],"   ", rn.list[[x]]))})
+  if (class(fix.all.unlist) == "character"){
+    fix.all.unlist = t(data.frame(fix.all.unlist))
+  }
   rownames(fix.all.unlist) = unlist(rn.list)
   tb.df = as.data.frame(rbind(fix.all.unlist, ranef.na))
   
@@ -113,12 +131,12 @@ lmer.display = function(lmerMod.obj, dec = 2, ci.ranef = F){
   ll = round(sl$logLik[[1]], dec)
   aic = round(sl$AICtab, dec)[1]
   met = c(NA, no.grp, no.obs, ll, aic)
-  met.mat = cbind(met, matrix(NA, length(met), 3))
+  met.mat = cbind(met, matrix(NA, length(met), ncol(fix.all) - 1 ))
   rownames(met.mat) = c("Metrics", paste("No. of groups (", rownames(met.mat)[2:(length(no.grp) + 1)], ")", sep=""), "No. of observations", "Log-likelihood", "AIC value")
   met.mat[, 1] = as.character(met.mat[, 1])
   colnames(met.mat) = colnames(tb.df)
   tb.lmerMod = rbind(tb.df, met.mat)
-  lapply(c(2,4), function(x){tb.lmerMod[, x] <<- as.numeric(as.vector(tb.lmerMod[, x]))})
+  lapply(seq(2, ncol(fix.all), by =2), function(x){tb.lmerMod[, x] <<- as.numeric(as.vector(tb.lmerMod[, x]))})
   
   ## caption
   cap.lmerMod = paste(sl$methTitle, " : ", y," ~ ", forms[3], sep="")