+++ NEWS.md
@@ -1,3 +1,9 @@
+## 0.4.0
+
+* Table from `coxme.object` (**coxme** package)
+
+* Error correct: 1 variable case.
+
 ## 0.3.5
 
 * Change default page length option of `opt.tb1` from 10 to 25.